see extended description
Full list of modifications

1. Boot screen still the same.

2. Change app package name

3. try another alarm module. check if time is behind now and add to tomorrow

4. The Testing found the incorrect word but also added the word after it as being incorrect.
a. It also highlighted GREAT below even though I didn’t type out the entire verse.

5. The text on both portrait and landscape should be bold, centered and auto size to fill the entire background.
a. We also don’t need “tap to show” or “hide”. This will be in the instructions.

6. Can we put a division between the 4 buttons or just make them a bit smaller in size?

7. Can we put icons up on the top so users don’t have to go back to the home screen to get to another location?
a. Change New Bible quotation to “Chapter & Verse”
b. Clean up the look a bit making “Select a Folder” a button and also making “SAVE” a button?
c. Would it be possible to show the folders that are already created (instead of having to click on “Select a Folder”) and the user could just click on the folder they want?

7. debug gsign in button